refactor: Route 구조 변경
