+++ src/App.tsx
@@ -6,7 +6,9 @@ import NotFound from "./pages/NotFound";
 import Profile from "./pages/Profile";
 import Signup from "./pages/Signup";
 import ResetPassword from "./pages/ResetPassword";
-import Dashboard from "./pages/dashboard";
+import DashboardHome from "./pages/dashboard/DashboardHome";
+import DashboardAllTeam from "./pages/dashboard/DashboardAllTeam";
+import DashboardTeam from "./pages/dashboard/DashboardTeam";
 import Plan from "./pages/Plan";
 
 function App() {
@@ -22,20 +24,12 @@ function App() {
         <Route path="/plan/:id" element={<Plan />} />
 
         {/* Dashboard routes wrapped with DashboardProvider */}
-        <Route
-          path="/dashboard/*"
-          element={
-            <DashboardProvider>
-              <Routes>
-                <Route path="/" element={<Dashboard />} />
-                <Route path="/home" element={<Dashboard />} />
-                <Route path="/all-team" element={<Dashboard />} />
-                <Route path="/shared-projects" element={<Dashboard />} />
-                <Route path="/team/:teamId" element={<Dashboard />} />
-              </Routes>
-            </DashboardProvider>
-          }
-        />
+        <Route path="/dashboard" element={<DashboardProvider />}>
+          <Route path="home" element={<DashboardHome />} />
+          <Route path="all-team" element={<DashboardAllTeam />} />
+          <Route path="shared-projects" element={<DashboardHome />} />
+          <Route path="team/:teamId" element={<DashboardTeam />} />
+        </Route>
 
         <Route path="*" element={<NotFound />} />
       </Routes>
