+++ src/pages/dashboard/DashboardAllTeam.tsx
@@ -1,13 +1,11 @@
 import DashboardHeader from "../../components/dashboard/DashboardHeader";
-import type { Project } from "../../types/project";
+import MainLayout from "../../components/layouts/MainLayout";
 import { Link } from "react-router-dom";
+import { useDashboards } from "../../contexts/DashboardContext";
 
 // 전체 팀 목록 페이지: 사용자가 속한 모든 팀을 보여주고 팀별 대시보드로 이동할 수 있는 링크 제공
-interface DashboardAllTeamProps {
-  projects: Project[];
-}
-
-const DashboardAllTeam = ({ projects }: DashboardAllTeamProps) => {
+const DashboardAllTeam = () => {
+  const { projects, loading, error } = useDashboards();
   const teams = Array.from(new Set(projects.map((p) => p.team.id)))
     .map((teamId) => {
       const project = projects.find((p) => p.team.id === teamId);
@@ -16,7 +14,7 @@ const DashboardAllTeam = ({ projects }: DashboardAllTeamProps) => {
     .filter((team): team is NonNullable<typeof team> => team != null);
 
   return (
-    <>
+    <MainLayout loading={loading} error={error}>
       <DashboardHeader title="전체 팀" />
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
         {teams.map((team) => (
@@ -29,7 +27,7 @@ const DashboardAllTeam = ({ projects }: DashboardAllTeamProps) => {
           </Link>
         ))}
       </div>
-    </>
+    </MainLayout>
   );
 };
